Shifted the Multi_processor.run() method up a level to Processor.run().
This completes one of the TODOs, and will be needed to avoid code duplication for handling the new
data_upload() and data_fetch() API methods.  This run() method now works for the uni-processor, as
the Uni_processor.on_master() method which always returns True has been reactived, so the
Uni_processor.run() method has been deleted.