The multi-processor package now allows sys.exit() calls within the master processor.
This affects all processor fabrics, as the Processor.run() method has been modified.  The SystemExit
exception is now specifically caught on the master.  The Processor.exit() method is then called to
allow the specific fabric to clean up, if needed, and then the SystemExit is re-raised to allow the
system exit to proceed as normal.