Bug fix for the GUI when deleting analysis tabs.
The deletion of analysis tabs was previously failing in certain cases.  The problem was the
generic_fns.pipes.delete() call in the middle of the delete_analysis() method causing a notification
to be sent to the pipe_deletion GUI observer object, which would then handle the pipe deletion prior
to the completion of the delete_analysis() call causing things to get out of sync.