ANSI escape characters are now turned off forcibly when in GUI mode.
The prompt.interpreter.Interpreter class was being initialised during the import startup phase due to
the gui.interpreter.Interpreter singleton class.  This meant that sys.stdout had not been redirected
to the relax controller yet and was still set to the terminal TTY.  This hack turns off the relax
prompt colouring forcibly.