The GUI tests now work in the GUI!!
This is thanks to a lot of GUI black magic.  The tests' tearDown() method now carefully deconstructs
the GUI element prior to the next test being run.  In the normal 'relax --gui-test' mode, the GUI
object is destroyed and recreated for each test.  However when run from the GUI, the GUI object is
always there and must remain intact.  Some of the careful deconstruction includes the deletion of
each analysis tab, the selective destruction of all non-main windows (excluding the controller which
shows the test suite progress).  The relax data store GUI object is also reconstructed in the
tearDown() method, and all wx events flushed at the very end to prevent clashes with the next GUI
test.