Improvements to the HTML version of the user manual.
The latex2html configuration script now allows for more maths in the HTML with the HTML_VERSION math
extension.  The section numbers are now removed from the HTML pages to allow for more static pages
for the user functions which don't disappear as new user functions are added, changing the section
numbering.  The title page has also been renamed to "The relax user manual" for better search engine
indexing.