Significant updates to the introduction chapter of the relax user manual.
The text has been updated to match the current relax versions.  All mention of the dead SRLS theory
has been eliminated.  Screenshots of all the GUI windows and analyses have been added to prettify
the chapter, and the text for the GUI has been rewritten for the new GUI design.