Created a new String_list wizard GUI element class and associated element_string_list() method.
This special object is currently very similar to the input_field() object, but is set up as a class
in the gui.wizard_elements module.  The wizard page element_string_list() method has been added as
the interface for adding this element to a wizard page.  This method stores the String_list element
in the wizard page _element dictionary.