Created the GUI test base method execute_uf() for executing a user function.
This is an interface which will massively simplify the creation of GUI tests, so that they can be
almost the same as normal relax scripts.  This will allow for the porting of most of the system test
scripts into GUI tests.  This however requires the use of the new wizard page element_*() methods
to create the GUI elements.