Creation of a branch for redesigning the internal structure of the user function.
The aim is to create a new package called 'user_functions' which will contain absolutely all
information about user functions, from documentation to icons to be used in the GUI.  This package
will contain a special data structure which will be used by the different UIs to automatically
generate their interfaces to the user functions.  This will significantly simplify the code base
and make it much easier to create user functions for all UIs simultaneously.