Big bug fix for the test suite!
One of the state system tests was removing the frame order analysis from the
generic_fns.pipes.VALID_TYPES list with a pop() call.  But this change is permanent.  So in the
later Bmrb.test_bmrb_rw GUI test, the GUI fails as this list is used in the pipe.create GUI user
function interface and it no longer matches the corresponding PIPE_DESC_LIST structure!