Converted the molmol user function to the auto-generation design.
For this process, the molmol.macro_apply and molmol.macro_write user function colour_start and
colour_end args had to be split, just as in the pymol user functions.  They previously accepted both
a Molmol or X11 colour name or an RGB colour array but are now separate args.