Converted the grace user functions to the auto-generation design.
This required created the generic_fns.grace.get_data_types() function for returning the data type
descriptions and values for the current data pipe.  This is to replace the code in the old grace
user function update_parameters() method so that the data type fields are updated according to the
analysis.