The user function args can now have minimum and maximum values for integers specified.
This was accidentally partly committed in the last commit!  The min and max args can now be given,
and these are then propagated all the way to the GUI elements.  For the Sequence and Sequence_2D
elements this has no effect yet as they do not implement wx.SpinCtrl elements.