relax version 1.3.16.
This version is mainly a documentation release.  The relax user manual changes include the expansion and completion of the model-free chapter of the relax user manual, addition of GUI screenshots and descriptions, a description of Gary Thompson's multi-processor framework, and general updates and improvements throughout.  The release also fixes the scons 'install' target for users installing relax in this way.  The Bruker Protein Dynamics Centre user function class has been renamed to 'bruker' because of Bruker's absorption of the PDC into the Bruker Dynamics Centre.  These user functions, as well as the BMRB user functions, have been added to the menu system.  Finally a test suite bug triggered in the Fink relax installation has been eliminated.