Fix for the Uf_container.add_keyarg() method for the wiz_combo_choices argument.
The argument defaulted to [] which is the problem!  Then for every single user function argument
which does not provide a value for wiz_combo_choices will all be aliased to the same object!  The
GUI code was then pre-pending the empty string to this list if can_be_none is set for the Value
object, modifying the list for all other user functions.