Changed the behaviour of the generate_spin_id() and generate_spin_id_data_array() functions.
This is a large change affecting many parts of the code base.  The problem that this solves is the
incorrect usage of the '&' character in spin IDs, tripping up on the powerful Selection object of
Chris MacRaild.

For example, the spin ID ':2&:Glu' cannot be correctly handled.  This actually matches all residues
with the number 2, and all residues with the name 'Glu'.  It is very strange that this has not been
noticed before!