Created a directory for GUI test 'scripts' to match the system test layout and scripts.
The idea is to mirror the system test script setup, at least for the scripts.  The GUI scripts will
be different by pushing the user functions through the execute_uf() method, but otherwise should
be identical to the system tests.  This will allow, in the future, possible code sharing for the
checking of the results of the script execution.