Refactored the system and GUI tests for code sharing.
The GUI tests now can inherit tests from the system tests.  This is done for the BMRB system tests,
both of which now run as GUI tests.  For this, a test class specific script_exec() method has been
added.  For the system tests, it passes the system tests script into the interpreter.  For the GUI
tests, it executes the script using the function locals so that the GUI test execute_uf() method is
accessible.