Bug fix for the calling of the GUI user function objects.
The user function page update_args() method is now being called to update all the ComboBoxes with
the current relax data store data.  Without this, the lists would be empty or out of date.  A call
requiring a user function arg to be set on a read only ComboBox would then fail.