Bug fix for all of the specific analysis code.
The SPIN_PARAM and GLOBAL_PARAM structures are now initialised within the __init__() methods rather
than being class variables.  This bug was only uncovered by the auto-generated user functions.  The
problem with them being class variables of the base class is that the objects are then the same for
all specific analyses.