Mac OS X bug fix for the non-read only ComboBox GUI input elements.
This is for the Value and Combo_list GUI input elements.  The problem was two-fold.  First the
convert_to_gui() call was required to convert the value to a unicode string, as a normal string will
be not be set via the ComboBox.SetValue() method.  Secondly old selections need to be turned off.
On other operating systems, the SetValue() call will reset this - but not on Macs!!!  Grrrrr......