Big fix for the auto-generated GUI user functions - the recycling of pages should now work.
The recycling of user function pages was failing due to the wizard _exec_count data structure.  A
click on the 'ok' button increments these counts, and a subsequent _ok() call was then suppressing
the execution of the user function.  The GUI tests were missing this, as the _ok() call was
bypassed by calling on_execute() instead.