Started to redesign the user function definition descriptions.
The desc object has been converted into a list which will contain instances of the new
Desc_container class.  This class will hold the description information as blocks of text, including
unformatted paragraphs, verbatim text, lists, tables, and relax prompt examples.  The object has a
number of methods used for inputting all of the information, as well as extracting it.

The information in the Desc_container object is not formatted, for example tables will be stored as
lists of lists.  This will remove the need for parsing list, tables, verbatim text, etc. by the
docs.latex.fetch_docstrings module.  But these will need to be recreated for the prompt help
strings, for the GUI user function descriptions, and for the user manual.