Created an algorithm for wrapping table columns to fit within status.text_width.
This if for the prompt user function docstring tables.  The columns are now dynamically wrapped so
that the table will not be wider than the given width.  For example the value.set user function
description table would normally have a total width of 238 characters, but this now fits within one
screen of text under all operating systems.