Created a new way of defining and inputting tables into the user function descriptions.
This introduces the user_functions.objects.Table for holding all the table data and metadata, and
the user_functions.data.Uf_tables singleton object for holding all the tables.  This design will
allow more metadata (for example using longtables in the LaTeX user manual) to be specified.  The
singleton collection of tables with unique labels will also allow the elimination of duplicated
tables in the PDF and HTML user manuals.  The new table title metadata will also be used for the
table headings in the user manuals.