The GUI user functions now use the new Desc_container and Table user function classes.
The descriptions for the user function wizard pages are now being taken from the new
user_functions.objects.Desc_container objects, lists of which are the user function definition
'desc' objects.  The tables via the user_functions.objects.Table object is being formatted as in the
prompt user function help text, though without vertical lines.  The lists and itemised lists are
currently formatted the same as in the prompt UI.