The state.load user function will now recreate the GUI tabs.
For this, the new state_load observer object has been created.  The GUI analysis load_from_store()
is registered with this object.  And to prevent fatal racing in the GUI, the state.load user
function operation is now synchronous.  The state.load and state.save file arg defaults are now also
set to 'state.bz2'.