The user function pages are now registered and unregistered with the 'gui_uf' observer object.
This allows the ComboBox args to be properly updated with asynchronous user function calls.  The old
update method suffered from racing and hence often updated before the user function had executed.
Now the on_display() method registers the GUI user function page update_args() method, and the
on_next() method unregisters it, allowing the observer to update the page only when it is visible to
the user.