Better handling of the GUI wizard termination.
The _handler_close() method is now handling the close events and together with the _ok() method,
both now execute the current page's on_next() method.  This means that on_next() is called for all
avenues for exiting the wizard.  This allows the user function pages to unregister themselves with
the observer objects.