Fix for the Mf.test_mf_auto_analysis() GUI test.
The recent changes registering the user function update_arg() method with the gui_uf observer object
means that direct calls to the user function page on_execute() method will no longer work.  The
wizard needs to be run to completion!  Therefore the run() or _ok() methods of the wizard holding
the page should be executed instead.