Improvement for the model-free auto-analysis in the GUI.
After execution, the main data pipe of the analysis tab is now set to the 'final' data pipe
generated by the dauvergne_protocol.  This means that when switching between analysis tabs, that
going back to the model-free tab will make the 'final' data pipe the current one, rather than the
data pipe used to initialise the analysis.