Added support for the user function arg type 'force flag' to the GUI user functions.
This arg type will, in the GUI, always return True.  This is because a user clicking on 'Ok' or
'Apply' is essentially saying that this operation should be performed.  The special Force_true class
was created for these arguments which simply ignores SetValue() calls and always returns True from
GetValue() calls.