Merged the relaxation data list GUI element with the base list element.
The base list element now has an Enable() method which simply enables or disables all buttons (as is
done in the relaxation data list).  The button size is now specified as a (x, y) tuple.  And the GUI
object can now be passed into the elements, as this is needed for the relaxation data metadata
window.