Refinement of all of the GUI user function wizard sizes.
The changes in this branch have significantly changed the description layouts.  Therefore all the
sizes were checked and reoptimised.  The sizes were optimised so that they are visible within a
1024x768 computer screen.  And the description heights together with wizard heights were optimised
to remove as much scrolling in the description as possible.

In the process, a few of the frame order prompt example descriptions were fixed.