Bug fix for sizing of the GUI about dialogs - the virtual sizes are now accurately calculated.
To do this, the window contents (a wx.BufferedDC instance) is created twice, with the virtual window
size exactly calculated in the first pass, and then used to set up the device context in the second.
A space is also now added to the end to ensure the border is there.