Started to fill up the spin ID look up table.
The index_molecule() and index_residue() functions have been added to determine the MoleculeList and
ResidueList indices of given molecules and residues.  These are used by the create_spin for
efficiency and to allow the indices (together with the spin index and spin ID string) to be
assembled into the look up table.  This table is not used anywhere yet.