Redesign of the setUp() and tearDown() methods of the GUI tests.
This is needed as the current design does not work under MS Windows!  Instead of created a new wx
app for each test (this dies in Windows as it looks like only a limited number of toolbars can be
created), only one app is created and is recycled.  This design better mimics GUI usage anyway.

This change uncovers a few GUI bugs not caught by the previous design.