relax version 2.1.0.
This major feature release introduces the new concept of a container for data linking two atoms or
spins, the interatomic data design.  This is for NMR data such as the magnetic dipole-dipole
interaction between two spins (including relaxation and RDC data) and NOESY data.  As this is a
highly invasive change, modifying almost all of the relax analysis types and modifying the GUI, the
minor version number has been incremented and all of the 2.0.x version numbering has been skipped.