Converted the frame order target function code to be more like the N-state model.
This matches a number of changes which occurred in the trunk (via branches).  The alignment index
has been changed from 'i' to 'align_index' for clarity.  The 'rdc_const' and 'pcs_atoms' keyword
arguments have been renamed to 'dip_const' and 'atomic_pos' respectively to match the N-state model.
And proper support for the interatomic data design for the RDCs has been added.