Redesigned the align_tensor.init user function.
This is to better handle multiple tensors per alignment (and/or per domain).  The tensor ID is now
optional and the alignment ID must be supplied.  The order of the arguments has also changed so all
the IDs (tensor, alignment, and domain) are at the start.  The function
generic_fns.align_tensor.get_tensor_object_from_align() has been added to handle the retrieval of
alignment tensors without a tensor ID or name.