relax version 2.1.1.
This release includes a major overhaul of the relax user manual with the addition of new chapters
and tutorials for all analyses.  It also includes a major set of bug fixes to re-enable the NOE, R1,
R2 and model-free analysis of multiple spin types (e.g. the Trp sidechain indole NH).