Fix for bug #20213 (https://gna.org/bugs/?20213).
The problem was that the generic_fns.sequence.generate() function was searching for a spin based on
its spin ID generated from the spin name and not spin number.  Hence two spins with the same name
but different numbers were treated as the same spin, and only data from one was being generated via
the copy() function.