Complete redesign of the frame order fully-automated protocol.
The protocol now uses a number of 'tricks' to allow the calculations to complete in a few days
rather than in approximately one million years!  These include:

- Sobol' quasi-random points for the numerical integration of the frame order equations for the PCS.

- Use of the parameters of simpler nested models as initial parameter estimates in the grid search
of the more advanced models to cut the number of dimensions down from 9 for the pseudo-ellipse to a
maximum of 3 for all the models.  For example the average domain position from the rigid model is
used as the starting point for all other models as this position is very close to the real average.

- Very low precision via a small number of Sobol' points for the initial grid search.  This is good
enough however to reach the grid search minimum.

- A zooming minimisation strategy whereby minimisation is repeated using higher and higher
precision, i.e. the number of Sobol' integration points increasing from 100, to 1000, to 200000.

- Initial solutions found using a PCS data subset.  Two sets of PCS data files are created.  The
first contains all PCS data.  The second contains a selection of ~5 reliable PCS values spread
across the whole domain.  The PCS subset is used for the initial grid search and for an initial
minimisation to massively speed up convergence to the minimum.