relax version 2.1.3.
This is a major bugfix release which marks the completion of the N-state model analysis code.  The
N-state model allows single static structures or ensembles of static structures to be analysed and
compared using residual dipolar couplings (RDCs), pseudo-contact shifts (PCSs), distance restraints
via NOEs.  In addition to alignment tensors, the populations or probabilities of each state can be
optimised as well as the position of the paramagnetic centre when PCSs are used.