Big set of fixes for the frame order analysis.
RDC and PCS data for the non-moving domain are now not used in the analysis.  This was not a problem
during development as the test models had no RDC and PCS data for the non-moving domain.  Now the
spin_loop() and interatom_loop() methods are called with the moving domain ID string as the
selection argument.  This avoids all non-moving domain data, as should happen for the analysis.