Fix for the frame order analysis for a bug introduced at r18566.
The _domain_moving() method has been redefined to return the spin ID string corresponding to the
moving domain rather than the domain ID, as this method is only used for restricting the spin_loop()
and interatomic_loop() methods.  The changes allows the loop methods to function prior to defining
the moving domain.