Initial attempt at allowing the average domain position to translate in the frame order analysis.
Previously only the domain orientation was optimised around a distant pivot point to achieve a
coupled translation and rotation.  This was, however, an artifact of history from when only RDCs
were being used in the analysis.  It should be possible to optimise both the translation and
orientation (independent of the pivot) of the moving domain.

For this, the frame_order.ave_pos_translation has been added to allow the translation optimisation
to be toggled.  The parameters 'ave_pos_x', 'ave_pos_y' and 'ave_pos_z' have been added to the
model.  The analysis specific code, auto-analysis, and target function code have been modified to
allow optimisation.  However there are some issues and the translation optimisation does not work.
But the original code paths do function.