The grid search for the frame order auto-analysis with translation is now handled differently.
This is to avoid getting trapped into a local minimum.  The grid for this split into a grid search
over the rotational space first, then a grid search over the translation.  This is repeated for good
measure.  This is much more accurate and reliable for finding the global minimum and is also much
faster.