Added a comment to the output from value.display and value.write to describe the parameter.
This idea is discussed at http://thread.gmane.org/gmane.science.nmr.relax.user/1428.

The idea is to take the parameter description from the specific analysis API and add it to the top
of the file or output.  This is to help understand what the Rex value are.  For example for the Rex
parameter the first line would be:

# Parameter description:  Chemical exchange relaxation (sigma_ex = Rex / omega**2).