Final synchronisation of all the frame_order_testing branch files back to trunk for the merger.
The manual merger of the frame_order_testing branch did not bring over all changes for some reason.
Maybe this was due to bad conflict resolution.  The changes not merged back were found with the
command:

$ diff -qr . ../../relax-trunk | grep -v ".svn"

This was run from the checked out branch.  All the differing files were then copied over manually.