Recreation of the relax_disp branch from the latest trunk code.
This branch will be a recreation of the old relax_disp branch which is no longer visible in the head
revision.  Each commit of the old branch will be manually ported with commands such as:

svn merge -r8242:8243 svn+ssh://bugman@svn.gna.org/svn/relax/branches/relax_disp .

The old branch will remain permanently in the repository, but will not be visible at head revision.