Fix for bug #20674 (https://gna.org/bugs/?20674) - the failure of the consistency testing analysis.
This was reported by Mengjun Xue <mengjun dott xue att mailbox dott tu-berlin dot de>.

The problem was that the first residue did not have a single proton 'H' in the PDB file, and
therefore the dipolar relaxation interaction was not set up.  The overfit_deselect() method of the
consistency testing specific API was not checking for this.  The method is now much more like that
of the model-free specific analysis.