Completely redesigned the minimisation parts of the specific_analyses.relax_disp module.
Instead of dealing with the optimisation of individual spins, groups of spins are now optimised
together.  This allows for the clustering analysis of relaxation dispersion.  The method
_block_loop() has been created to loop over spin blocks, but it currently only returns individual
spins.  However with the rest of the code designed to handle this loop, only this function needs
to be modified to enable clustering.

The method _param_num() has also been added to determine the total parameter number per spin block.
The data structures sent into the Dispersion target function class have also been redesigned to
handle spin blocks instead of individual spins.