The 'exp_fit' relaxation dispersion model now uses the minfx.grid sparseness argument.
This is used to skip all parts of the grid search belonging to a different exponential curve or
different spin.  If the number of curves is N and the number of spins M, the grid size decreases
from inc**(2*N*M) to (inc**2)*N*M.  For lots of spins and curves, this is a huge decrease.