The relaxation dispersion target function class back_calc variable now matches the values variable.
Instead of being a temporary structure which is overwritten for each spin and each exponential
curve, the structure now matches the dimensions of the values variable and hence is persistent per
function call.  This allows external code to access the structure - for example for data back
calculation in the relaxation dispersion specific analysis module.