Modified pipe_control.plotting.assemble_data_series_series() to handle dictionaries with keys as values.
This will be useful in, for example, relaxation dispersion for plotting the dispersion curves.  In
this case, the R2eff values are in a dictionary where the keys are the values to plot against.  This
is different from the current case where the X and Y data dictionaries are required to have the same
keys.  These changes expand the capabilities of the grace.write user function.