The relax_disp.exp_type user function has been shifted to the new analysis wizard.
Instead of being one of the elements on the relaxation dispersion analysis frame, it is now placed
between the analysis selection page and the data pipe page of the new analysis wizard.  The user
function execution is delayed until the set up of the frame, just after the execution of the
pipe.create user function.  This will allow the frame to be set up differently for each experiment
type.