Elimination of the relax_disp.calc_r2eff user function.
This user function, which is non-functional anyway, is not needed.  The calculation of the R2eff
values will occur with the optimisation of the 'R2eff' model (with a call to the calc user function
for the fixed time period experiment types) so direct calculation through a specific user function
is not needed.