Removed some bad calls to status.exec_lock.release().
This commit may have to be reverted in the future.  The problem is that the execution lock is not
being held when these calls are made.  The calls were added as a kludge to handle certain situations
where the execution lock was not released.  There may be cases were this behaviour is still needed.