Big bug fix for a memory leak in the relaxation curve-fitting C module.
Proper reference counting is used for the temporary 'element' Python objects used in the conversion
between Python and C objects.  The use of the Py_CLEAR() macro removes the memory leak.  However the
number of references as seen by sys.gettotalrefcount() in a debugging Python version keeps rising
and might be a problem in the future.