Created the Relax_disp.test_hansen_cpmgfit_input system test.
This is for checking the operation of the relax_disp.cpmgfit_input user function conversion of
Flemming Hansen's CPMG R2eff values into input files for CPMGFit.  A relax state file containing the
results of an analysis of an 'R2eff' model analysis of the truncated data has been added to the test
suite data to speed up the test and to check the loading of dispersion state files.