Added the results from NESSY of the analysis of Flemming Hansen's truncated CPMG data.
This is only for the truncated data of residues 70 and 71.  All files, except for the PNG graphics,
have been added to the repository.  The 'summary' file has been created to hold the data from
NESSY's summary tab, as this is not stored in the NESSY saved state and is permanently lost after
closing NESSY.