Added support for setting the spin isotope information in the dispersion GUI.
A new Text_ctrl element has been added just after the spin system GUI element.  This displays a list
of all the spin isotopes currently defined and is updated after every GUI user function call.  The
button of the element launches the spin.isotope user function.

The spin isotope information is now checked for prior to executing the GUI analysis and added to the
missing list to present to the user when blocking the execution of the analysis.

The dispersion GUI test has been updated to use this new element.