Completely redesigned how parameters are handled in the relaxation dispersion analyses.
The key concept is that everything revolves around the new loop_parameter() function.  This is a
generator function which loops over the parameters of a given cluster, yielding all the information
required to access the parameter.  The other functions of the parameters module use loop_parameter()
to sequentially handle each parameter.  This allows for huge simplifications of these functions.