Fix for bug #20888 (https://gna.org/bugs/?20888), the autoscaling of Grace graphs.
This solution was mentioned in the post at
http://thread.gmane.org/gmane.science.nmr.relax.devel/3920/focus=3930.

Instead of using minimum and maximum values for the axes in the Grace graphs produced by the
lib.software.grace module, which was the old solution for having the graphs scaled to reasonable
values, instead the '@autoscale' command is appended to the end of all graphs.  This is performed by
the write_xy_data() function.