Completed the fix for bug #20889 (https://gna.org/bugs/?20889).
The problem was that the spectrum.read_intensities user function was incorrectly updating the
cdp.spectrum_ids list when the spectrum_id argument is set to a list.  The list of IDs was being set
as a single element of cdp.spectrum_ids, causing problems with the GUI when updating the ComboBox
choices and then subsequent setting of the spectrum IDs.

This bug and fix is independent of the relax_disp branch, despite being uncovered there and being
caught by the Relax_disp.test_bug_20889_multi_col_peak_list GUI test in that branch.