Fix for bug #20915, (https://gna.org/bugs/?20915) - Failure of Grace opening in MS Windows
Troels E. Linnet provided this patch, and was discovered during work on a Windows 7 system: telinnet
aaattt bio_dot_ku_dot_dk

This is a small fix for a wrong call to "raise RelaxMissingBinaryError(binary)", when issuing an
external call to xmgrace.

The "path_sep" would be equal = [\/], and the RE search would not find(True) the full path specified
for the xmgrace file.  This is now shifted to python: os.path.isfile
http://docs.python.org/2/library/os.path.html

Another fix, is that as a standard the command "xmgrace" is provided. This will work fine through
windows cmd, but the true name for program in windows is "xmgrace.exe", and so an additional search
for +".exe" is also performed.