Fix for bug #20916, (https://gna.org/bugs/?20916) - Suggestion for python script for PNG/EPS/SVG conversion of grace files.
Troels Linnet provided this patch, and was discovered during work on a Windows 7 system: telinnet
aaattt bio_dot_ku_dot _dk

This patch will provide a grace2images.py file in each folder where a call to
specific_analyses/relax_disp/disp_data.py is called.

It is called in: def plot_disp_curves(dir=None, force=None):
and call the function: lib.software.grace.script_grace2images()

The conversion script can be executed in linux and windows, if the PATH to xmgrace has been
specified. It will look in a folder for grace files of ending *.agr and by default convert to PNG.

One can also convert to EPS and SVG. Propably more options could be added, as PDF. The conversion
depends on xmgrace compilation, and so PNG conversion is for fast inspection of graphs in folder,
and EPS for further external conversion to PDF etc.

The patch, the output file, and small script to test is attached.

I miss to make the file executable in relax, so the script can be executed directly in Linux.