Fix for bug #20917 (https://gna.org/bugs/?20917).
The problem is that the Grace files for each spin system are not created by the
relax_disp.plot_disp_curves user function as the ':' character cannot be placed in a file name in MS
Windows.  All of the file name from the ':' onwards is lost.

The solution is to replace each of the characters '#:@' in the spin ID string with '_'.