More MS Windows fixes, this time a nasty kludge, for the relax system tests.
This is strangely needed for the relax_disp branch and not the trunk for a 64-bit MS Windows 7 test
system.  The reason why this WindowsError is triggered by the base tearDown() method in the
relax_disp branch and not trunk is a total mystery.  Actually why Windows refuses to complete the
file close() operations of the results.write and state.save user functions before calling the
tearDown() method is the greater mystery.