Improvement of the error handling in the 'NS 2-site star' model.
The fA and pB parameters are no longer being checked.  Instead a Mgx value of 0.0 is being checked
for.  This catches additional problems.  And now instead of the R2eff value being set to zero, it is
set to 1e99.  This is because log of zero is -inf, and then multiplied by a negative constant gives
positive inf.