Added the 'TP02' R1rho' calculating function to the relax library.
This is the Trott and Palmer 2002 R1rho analytic model for 2-site exchange.

This commit follows step 3 of the relaxation dispersion model addition tutorial
(http://thread.gmane.org/gmane.science.nmr.relax.devel/3907).  The Matlab code from Skrynnikov and
Tollinger has not been converted to Python code yet.  This is to allow the Matlab->Python conversion
to be followed.