An attempt at better handling MS Windows not releasing the file handle on time in the test suite.
The system and unit tests tearDown() method should now be resilient to the strange MS Windows
behaviour of not releasing the relax state files.  The tearDown() method should now complete even
when this error occurs.  A delay of 3 seconds has been added when the WindowsError occurs to give
the OS some time before attempting to delete the file again.  If this fails, then the file deletion
operation is skipped.