relax version 3.0.0.
This is the first release of the new relax 3 series.  This release marks a major shift of relax towards becoming a scientific computing environment designed for the study of molecular dynamics using experimental biophysical data.  It is designed to be a replacement for numerical computational environments such as GNU Octave, MATLAB, Mathematica, Maple, etc.  From the perspective of a user, however, not much has changed.  There are no major modifications to the prompt, script, or graphical user interfaces.  The changes are rather in the backend.  The infrastructure changes are comprehensive and include the renaming of most of the relax packages and the shift of huge quantities of code from all parts of the code base into standalone modules in the relax library.