Added support for the Mac OS X Modelfree4 binary results to the Palmer.* system tests.
The Mac OS X Modelfree 4.20 binary produces different results than the Linux binaries, mainly due to
a compilation problem.  In the Linux binaries, the results are written out to 4 decimal places.  In
the Mac binaries, the results are instead written out to 4 significant figures.  Therefore the
number of decimal places are much less than the Linux results.