Attempt at fixing bug #21080 (https://gna.org/bugs/?21080).
This was reported by Troels Linnet.

The problem is a standard GUI problem.  The text from a wxPython GUI is a unicode string.  But relax
requires standard strings.  Therefore the gui.string_conv.gui_to_str() function needs to be used on
the return value of the GUI element GetValue() method, but this call was forgotten.