The tutorial for adding dispersion models in the user manual has been simplified.
Most of the text from the dispersion model addition tutorial in the dispersion chapter of the manual
has been removed.  Instead a link to the tutorial on the wiki is given as this is a much better
place for such information
(http://wiki.nmr-relax.com/Tutorial_for_adding_relaxation_dispersion_models_to_relax).