Python 3 fixes for all of the unicode strings in relax.
Instead of using the u"xyz" notation, now unicode("xyz") is being used.  This works as the relax
compat module sets the builtin unicode() function to str() for Python 3, as all strings in Python 3
are unicode and hence both the Python 2 u"xyz" and unicode() code are undefined in Python 3.