Huge speed up of the GUI tests by the removal of the N_state_model.test_populations test.
This problem was identified by running the GUI tests with the '--time' flag.  One one test machine,
this single test took ~142 seconds to complete when the entire GUI tests took ~242 seconds (i.e.
this one test took up to 60% of the whole test suite).  This test comes directly from a system test,
but the equivalent system test only takes about 6 seconds to complete.  The difference is due to
the slow generation of the user function GUI pages.