Redesign of the model list GUI element.
The wx.ListCtrl element has been replaced by a wx.FlexGridSizer combined with wx.CheckBox and
wx.StaticText.  The result is a much nicer formatting of the element.  The checkboxes in the old
element displayed slight rendering problems on all operating systems and did not look neat.  The new
design is also more flexible in that models of None are now treated as separators in the window.