Huge speed win for the relaxation dispersion analysis - optimisation now uses the multi-processor.
The relaxation dispersion optimisation has been parallelised at the level of the spin clustering.
It uses Gary Thompson's multi-processor framework.  This allows the code to run on multi-core, multi
-processor systems, clusters, grids, and anywhere the OpenMPI protocol is available.

Because the parallelisation is at the cluster level there are some situations, whereby instead of
optimisation being faster when running on multiple slaves, the optimisation will be slower.  This is
the case when all spins being studied in clustered into a small number of clusters.  It is also
likely to be slower for the minimise user function when no clustering is defined, due to the
overhead costs of data transfer (but for the numeric models, in this case there will be a clear
win).

The two situations where there will be a huge performance win is the grid_search user function when
no clustering is defined and the Monte Carlo simulations for error analysis.