Changed the operation of the grid search for the relaxation dispersion analysis.
If a parameter is a simple floating number type and it already has a value, then the grid search
over that dimension is fixed.  The grid increments are set to 1, and the upper and lower bounds set
to the parameter value.  This allows parameters to be pre-set, if known from experiment.  They will
nevertheless be optimised via the minimise user function.