A bit of help for some of the R1rho dispersion model system tests.
These fail after the fundamental fix of r21093.  The problem is only due to the very coarse grid
search size - a finer grid search allows the solution to be correctly found.  However as this is far
too slow, instead the kex parameter is set to be close to the solution to skip a grid search
dimension.