The dispersion multi-processor optimisation code now prints out its own simulation messages.
This is to fix bug #21190 (https://gna.org/bugs/?21190).

The memo object now is feed in the spin IDs of the cluster and stores this as the cluster_name
variable.  This is used by the results object run() method, which is run on the master at the end,
to print out a message along the lines of "Simulation X, cluster yyy".  Therefore the message is
only printed out once the calculation of that slave command is complete and returned to the master.