Replaced all usage of scipy.linalg.expm() with lib.linear_algebra.matrix_exponential.matrix_exponential().
This is for the functions of the lib.dispersion package used for the relaxation dispersion numeric
solution models.  The change eliminates a bug in the scipy function which uses the Pade
approximation which fails horribly for the complex part of the matrix.  The real part looks good,
but the complex part looks to have nasty truncation artefacts which is propagated and amplified
through the Bloch-McConnell equations.