Converted the Relax_disp.test_sprangers_cpmg_data_auto_analysis system test to not use the auto-analysis.
The test has been renamed to Relax_disp.test_sprangers_data_to_mq_ns_cpmg_2site.  The optimisation
is now for the cluster and has been severely cut back.  The 'MQ NS CPMG 2-site' model appears to be
rubbish anyway - it looks to be indeterminate with multiple solutions, and possibly infinite lines
of solutions.  The test now passes, and quickly.