Added a relax script for the 'MQ CR72' model optimised using Flemming Hansen's CPMG data.
This is to demonstrate, though not exactly successfully, that the 'MQ CR72' model can collapse to
the 'CR72' model.  The imperfection might be due to truncation artefacts in the sin ratio in the mD
and mZ factors.

The results files and output log file from the script have been added to the repository as well.