Redesigned the Relax_disp.test_hansen_cpmg_data_to_ns_cpmg_2site_star_full system test.
The R2A and R2B rates cannot be distinguished for this data, therefore there was no unique solution.
This resulted in too much variability between 32 and 64-bit systems as well as different operating
systems.  Instead a single calc user function call is used to determine the chi-squared value for a
fixed set of parameters.