Conversion of the format of the relaxation dispersion R2eff/R1rho data structures.
These are now lists of lists of lists of numpy arrays instead of pure numpy rank-4 arrays.  This
only affects a number of related data structures in the dispersion target function class.  The main
purpose is to prepare to have a different number of dispersion points per experiment, per spin, and
per spectrometer frequency.