Added the 'MP05' model to the relax user manual.
This is the Miloushev and Palmer 2005 R1rho analytic model for 2-site off-resonance exchange.

This follows the tutorial for adding relaxation dispersion models at:
http://wiki.nmr-relax.com/Tutorial_for_adding_relaxation_dispersion_models_to_relax#The_relax_manual.

The 'MP05' model was already partly in the manual, however it was listed as unimplemented.  All of
the tables and the dispersion chapter text has been updated for the model.