The 'MP05' and 'NS R1rho 2-site' are now nested in the dispersion auto-analysis.
As the 'MP05' model is valid across all times scales and does not require skewed populations, its
optimised parameters can be used as the starting point of optimisation of the 'NS R1rho 2-site'
numeric model.  This results in huge speed ups of the numeric model as previously a grid search was
being performed.